barely anything

+++ app/src/main/java/com/example/assignment2/ui/clinic/EditPatientActivity.java
@@ -138,6 +138,6 @@ public class EditPatientActivity extends AppCompatActivity implements PersonData
         TextView icText = findViewById(R.id.icText);
         icText.setText(icNumber);
 
-        setRadioButtons(db.getCurrentUser().getVaccineStatus());
+        setRadioButtons((int) db.getCurrentUser().getVaccineStatus());
     }
 }
